SCC-2207 Add holding class for retrieval
To integrate Holdings data into the discovery API, we need to fetch the statements for the holdings separate from the item statements.

The query SQL query fetches simply an array of statements associated with a bib (or multiple bibs) record at uses this to construct the item objects. I think it would be possible to modify that query (possibly with a `FILTER` in the `item_statement` aggregation), but I think that might be difficult to maintain and understand in the future. This simply checks the identifier prefix for each statement returned in the `item_statements` array and sorts them on if it is an `i` or `h`. This is a simple approach that I don't think poses any performance issues since it is such a simple check.

Some questions I have on this approach:

- Should we rename the `item_statement` in the query? I've left it as is for now since I couldn't think of anything better
- Is there any potential for side effects here? I've tested the code and I believe it will just pass through item statements if no holding statements are present.
- Philosophically, is this better to be done in the query? I could see an argument for that, but I actually worry it would have a greater performance impact since it would involve adding another subquery, though in likelihood, that would have less of an impact that running `filter` on the existing array.