Add support for loading index property in statements
To ease testing changes to the discovery-api-indexer, which uses this
component, we need discovery-store-models models to load the `index`
property in fixtures. This will allow us to better emulate the data that
the discovery-store-poster will make available to the
discovery-api-indexer. We haven't previously used the index property
when indexing values in Elasticsearch, but recent parallels work does
require it.

Related:

https://jira.nypl.org/browse/SCC-3039