Remove statement sorting (i.e. don't override sort)
Seems there was code to sort statements for a given predicate by
object_id or literal length. This was probably done partially to coerce
a somewhat consistent sort (of object_ids) and improve display (shorter
titles should probably be shown before longer titles?) In any case, it
was overriding the default sort the statements had when extracted in
some cases, which is throwing off parallel field relationships. So let's
not set a custom sort and just accept the original order of statements.

This is v1.4.2